Wed, Apr 29, 2026, 7:15 AM +02:00

+++ Readability.c
@@ -29,6 +29,6 @@ int main(void)
 
    int NivelDeGrado = 0.0588 * L - 0.296 * S - 15.8;
 
-   printf("%i\n", NivelDeGrado);
+   printf("%f\n", (double) NivelDeGrado);
 
 }
